+++ spec/rubocop/cop/rails/time_zone_assignment_spec.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+RSpec.describe RuboCop::Cop::Rails::TimeZoneAssignment, :config do
+  it 'registers an offense for `Time.zone=`' do
+    expect_offense(<<~RUBY)
+      Time.zone = 'EST'
+      ^^^^^^^^^^^^^^^^^ Use `Time.use_zone` with blocks instead of `Time.zone=`.
+    RUBY
+  end
+
+  it 'accepts `Time.use_zone`' do
+    expect_no_offenses(<<~RUBY)
+      Time.use_zone('EST') do
+        current_est_time = Time.zone.now
+      end
+    RUBY
+  end
+end