+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New features
 
 * [#446](https://github.com/rubocop/rubocop-rails/issues/446): Add new `Rails/RequireDependency` cop. ([@tubaxenor][])
+* [#458](https://github.com/rubocop/rubocop-rails/issues/458): Add new `Rails/TimeZoneAssignment` cop. ([@olivierbuffon][])
 
 ### Bug fixes
 
@@ -350,3 +351,4 @@
 [@flanger001]: https://github.com/flanger001
 [@ohbarye]: https://github.com/ohbarye
 [@tubaxenor]: https://github.com/tubaxenor
+[@olivierbuffon]: https://github.com/olivierbuffon